+++ src/alpha_mul_div.rs
@@ -1,11 +1,13 @@
+use std::ops::Deref;
 use std::sync::{Arc, Mutex};
-
-use fast_image_resize as fir;
-use pyo3::prelude::*;
 
 use crate::image_view::Image;
 use crate::pil_image_wrapper::{PilImageWrapper, RgbMode};
+use crate::thread_pool::ResizerThreadPool;
 use crate::utils::{cpu_extensions_from_u8, cpu_extensions_to_u8, result2pyresult};
+use fast_image_resize as fir;
+use pyo3::prelude::*;
+use pyo3::types::PyInt;
 
 #[pyclass]
 pub struct RustAlphaMulDiv {
@@ -22,16 +24,16 @@ impl RustAlphaMulDiv {
     }
 
     /// Returns CPU extensions.
-    #[pyo3(text_signature = "($self)")]
-    fn get_cpu_extensions(&self, py: Python) -> PyResult<PyObject> {
+    #[pyo3(signature = ())]
+    fn get_cpu_extensions<'py>(&self, py: Python<'py>) -> PyResult<Bound<'py, PyInt>> {
         let mul_div_mutex = self.mul_div.clone();
         let mul_div = result2pyresult(mul_div_mutex.lock())?;
         let cpu_extensions = cpu_extensions_to_u8(mul_div.cpu_extensions());
-        Ok(cpu_extensions.into_py(py))
+        Ok(cpu_extensions.into_pyobject(py)?)
     }
 
     /// Set CPU extensions.
-    #[pyo3(text_signature = "($self, extensions)")]
+    #[pyo3(signature = (extensions))]
     fn set_cpu_extensions(&mut self, extensions: u8) -> PyResult<()> {
         let cpu_extensions = cpu_extensions_from_u8(extensions);
         let mul_div_mutex = self.mul_div.clone();
@@ -45,25 +47,48 @@ impl RustAlphaMulDiv {
     /// Multiplies RGB-channels of source image by alpha-channel and store
     /// result into destination image.
     /// The images are represented as instances of ImageView.
-    #[pyo3(text_signature = "($self, src_image, dst_image)")]
-    fn multiply_alpha(&self, py: Python, src_image: &Image, dst_image: &mut Image) -> PyResult<()> {
+    #[pyo3(signature = (src_image, dst_image, thread_pool=None))]
+    fn multiply_alpha(
+        &self,
+        py: Python,
+        src_image: &Image,
+        dst_image: &mut Image,
+        thread_pool: Option<ResizerThreadPool>,
+    ) -> PyResult<()> {
         let mul_div_mutex = self.mul_div.clone();
         py.allow_threads(move || {
             let src_image_view = src_image.src_image_view();
             let dst_image_view = dst_image.dst_image_view();
-            let mul_div = result2pyresult(mul_div_mutex.lock())?;
+            let mul_div_guard = result2pyresult(mul_div_mutex.lock())?;
+            let mul_div = mul_div_guard.deref();
+            if let Some(thread_pool) = thread_pool {
+                return thread_pool.run_within(|| {
+                    result2pyresult(mul_div.multiply_alpha(src_image_view, dst_image_view))
+                });
+            }
             result2pyresult(mul_div.multiply_alpha(src_image_view, dst_image_view))
         })
     }
 
     /// Multiplies RGB-channels of image by alpha-channel inplace.
     /// The image is represented as instances of ImageView.
-    #[pyo3(text_signature = "($self, image)")]
-    fn multiply_alpha_inplace(&self, py: Python, image: &mut Image) -> PyResult<()> {
+    #[pyo3(signature = (image, thread_pool=None))]
+    fn multiply_alpha_inplace(
+        &self,
+        py: Python,
+        image: &mut Image,
+        thread_pool: Option<ResizerThreadPool>,
+    ) -> PyResult<()> {
         let mul_div_mutex = self.mul_div.clone();
         py.allow_threads(move || {
             let dst_image_view = image.dst_image_view();
-            let mul_div = result2pyresult(mul_div_mutex.lock())?;
+            let mul_div_guard = result2pyresult(mul_div_mutex.lock())?;
+            let mul_div = mul_div_guard.deref();
+            if let Some(thread_pool) = thread_pool {
+                return thread_pool.run_within(|| {
+                    result2pyresult(mul_div.multiply_alpha_inplace(dst_image_view))
+                });
+            }
             result2pyresult(mul_div.multiply_alpha_inplace(dst_image_view))
         })
     }
@@ -71,12 +96,13 @@ impl RustAlphaMulDiv {
     /// Multiplies RGB-channels of source image by alpha-channel and store
     /// result into destination image.
     /// The images are represented as instances of PilImageView.
-    #[pyo3(text_signature = "($self, src_image, dst_image)")]
+    #[pyo3(signature = (src_image, dst_image, thread_pool=None))]
     fn multiply_alpha_pil(
         &self,
         py: Python,
         src_image: &PilImageWrapper,
         dst_image: &mut PilImageWrapper,
+        thread_pool: Option<ResizerThreadPool>,
     ) -> PyResult<()> {
         if !src_image.is_rgb_mode(py)? {
             return result2pyresult(Err("Invalid mode of source PIL image"));
@@ -86,22 +112,37 @@ impl RustAlphaMulDiv {
         }
         let mul_div_mutex = self.mul_div.clone();
         py.allow_threads(move || {
-            let mul_div = result2pyresult(mul_div_mutex.lock())?;
+            let mul_div_guard = result2pyresult(mul_div_mutex.lock())?;
+            let mul_div = mul_div_guard.deref();
+            if let Some(thread_pool) = thread_pool {
+                return thread_pool
+                    .run_within(|| result2pyresult(mul_div.multiply_alpha(src_image, dst_image)));
+            }
             result2pyresult(mul_div.multiply_alpha(src_image, dst_image))
         })
     }
 
     /// Multiplies RGB-channels of image by alpha-channel inplace.
     /// The image is represented as instance of PilImageView.
-    #[pyo3(text_signature = "($self, image)")]
-    fn multiply_alpha_pil_inplace(&self, py: Python, image: &mut PilImageWrapper) -> PyResult<()> {
+    #[pyo3(signature = (image, thread_pool=None))]
+    fn multiply_alpha_pil_inplace(
+        &self,
+        py: Python,
+        image: &mut PilImageWrapper,
+        thread_pool: Option<ResizerThreadPool>,
+    ) -> PyResult<()> {
         if !image.is_rgb_mode(py)? {
             return result2pyresult(Err("Invalid mode of PIL image"));
         }
         let mul_div_mutex = self.mul_div.clone();
         py.allow_threads(|| {
-            let mul_div = result2pyresult(mul_div_mutex.lock())?;
-            result2pyresult(mul_div.multiply_alpha_inplace(image))
+            let mul_div_guard = result2pyresult(mul_div_mutex.lock())?;
+            let mul_div = mul_div_guard.deref();
+            if let Some(thread_pool) = thread_pool {
+                thread_pool.run_within(|| result2pyresult(mul_div.multiply_alpha_inplace(image)))
+            } else {
+                result2pyresult(mul_div.multiply_alpha_inplace(image))
+            }
         })?;
         image.set_rgb_mode(py, RgbMode::Rgba)
     }
@@ -109,25 +150,47 @@ impl RustAlphaMulDiv {
     /// Divides RGB-channels of source image by alpha-channel and store
     /// result into destination image.
     /// The images are represented as instances of ImageView.
-    #[pyo3(text_signature = "($self, src_image, dst_image)")]
-    fn divide_alpha(&self, py: Python, src_image: &Image, dst_image: &mut Image) -> PyResult<()> {
+    #[pyo3(signature = (src_image, dst_image, thread_pool=None))]
+    fn divide_alpha(
+        &self,
+        py: Python,
+        src_image: &Image,
+        dst_image: &mut Image,
+        thread_pool: Option<ResizerThreadPool>,
+    ) -> PyResult<()> {
         let mul_div_mutex = self.mul_div.clone();
         py.allow_threads(move || {
             let src_image_view = src_image.src_image_view();
             let dst_image_view = dst_image.dst_image_view();
-            let mul_div = result2pyresult(mul_div_mutex.lock())?;
+            let mul_div_guard = result2pyresult(mul_div_mutex.lock())?;
+            let mul_div = mul_div_guard.deref();
+            if let Some(thread_pool) = thread_pool {
+                return thread_pool.run_within(|| {
+                    result2pyresult(mul_div.divide_alpha(src_image_view, dst_image_view))
+                });
+            }
             result2pyresult(mul_div.divide_alpha(src_image_view, dst_image_view))
         })
     }
 
     /// Divides RGB-channels of image by alpha-channel inplace.
     /// The image is represented as instances of ImageView.
-    #[pyo3(text_signature = "($self, image)")]
-    fn divide_alpha_inplace(&self, py: Python, image: &mut Image) -> PyResult<()> {
+    #[pyo3(signature = (image, thread_pool=None))]
+    fn divide_alpha_inplace(
+        &self,
+        py: Python,
+        image: &mut Image,
+        thread_pool: Option<ResizerThreadPool>,
+    ) -> PyResult<()> {
         let mul_div_mutex = self.mul_div.clone();
         py.allow_threads(move || {
             let dst_image_view = image.dst_image_view();
-            let mul_div = result2pyresult(mul_div_mutex.lock())?;
+            let mul_div_guard = result2pyresult(mul_div_mutex.lock())?;
+            let mul_div = mul_div_guard.deref();
+            if let Some(thread_pool) = thread_pool {
+                return thread_pool
+                    .run_within(|| result2pyresult(mul_div.divide_alpha_inplace(dst_image_view)));
+            }
             result2pyresult(mul_div.divide_alpha_inplace(dst_image_view))
         })
     }
@@ -135,12 +198,13 @@ impl RustAlphaMulDiv {
     /// Divides RGB-channels of source image by alpha-channel and store
     /// result into destination image.
     /// The images are represented as instances of PilImageView.
-    #[pyo3(text_signature = "($self, src_image, dst_image)")]
+    #[pyo3(signature = (src_image, dst_image, thread_pool=None))]
     fn divide_alpha_pil(
         &self,
         py: Python,
         src_image: &PilImageWrapper,
         dst_image: &mut PilImageWrapper,
+        thread_pool: Option<ResizerThreadPool>,
     ) -> PyResult<()> {
         if !src_image.is_rgb_mode(py)? {
             return result2pyresult(Err("Invalid mode of source PIL image"));
@@ -150,22 +214,37 @@ impl RustAlphaMulDiv {
         }
         let mul_div_mutex = self.mul_div.clone();
         py.allow_threads(move || {
-            let mul_div = result2pyresult(mul_div_mutex.lock())?;
+            let mul_div_guard = result2pyresult(mul_div_mutex.lock())?;
+            let mul_div = mul_div_guard.deref();
+            if let Some(thread_pool) = thread_pool {
+                return thread_pool
+                    .run_within(|| result2pyresult(mul_div.divide_alpha(src_image, dst_image)));
+            }
             result2pyresult(mul_div.divide_alpha(src_image, dst_image))
         })
     }
 
     /// Divides RGB-channels of image by alpha-channel inplace.
     /// The image is represented as instance of PilImageView.
-    #[pyo3(text_signature = "($self, image)")]
-    fn divide_alpha_pil_inplace(&self, py: Python, image: &mut PilImageWrapper) -> PyResult<()> {
+    #[pyo3(signature = (image, thread_pool=None))]
+    fn divide_alpha_pil_inplace(
+        &self,
+        py: Python,
+        image: &mut PilImageWrapper,
+        thread_pool: Option<ResizerThreadPool>,
+    ) -> PyResult<()> {
         if !image.is_rgb_mode(py)? {
             return result2pyresult(Err("Invalid mode of PIL image"));
         }
         let mul_div_mutex = self.mul_div.clone();
         py.allow_threads(|| {
-            let mul_div = result2pyresult(mul_div_mutex.lock())?;
-            result2pyresult(mul_div.divide_alpha_inplace(image))
+            let mul_div_guard = result2pyresult(mul_div_mutex.lock())?;
+            let mul_div = mul_div_guard.deref();
+            if let Some(thread_pool) = thread_pool {
+                thread_pool.run_within(|| result2pyresult(mul_div.divide_alpha_inplace(image)))
+            } else {
+                result2pyresult(mul_div.divide_alpha_inplace(image))
+            }
         })?;
         image.set_rgb_mode(py, RgbMode::RgbA)
     }