+++ src/resizer.rs
@@ -1,12 +1,13 @@
+use std::ops::DerefMut;
 use std::sync::{Arc, Mutex};
-
-use fast_image_resize as fr;
-use fast_image_resize::{ResizeOptions, SrcCropping};
-use pyo3::prelude::*;
 
 use crate::image_view::Image;
 use crate::pil_image_wrapper::PilImageWrapper;
+use crate::thread_pool::ResizerThreadPool;
 use crate::utils::{cpu_extensions_from_u8, cpu_extensions_to_u8, result2pyresult};
+use fast_image_resize as fr;
+use pyo3::prelude::*;
+use pyo3::types::PyInt;
 
 fn filter_type_from_u8(filter: u8) -> fr::FilterType {
     match filter {
@@ -33,17 +34,26 @@ fn filter_type_as_u8(filter_type: fr::FilterType) -> u8 {
 }
 
 #[pyclass]
-pub struct RustResizeOptions(ResizeOptions);
+pub struct RustResizeOptions {
+    fir_options: fr::ResizeOptions,
+    thread_pool: Option<ResizerThreadPool>,
+}
 
 #[pymethods]
 impl RustResizeOptions {
     #[new]
     fn new() -> Self {
-        Self(ResizeOptions::new())
+        Self {
+            fir_options: fr::ResizeOptions::new(),
+            thread_pool: None,
+        }
     }
 
     fn copy(&self) -> Self {
-        Self(self.0)
+        Self {
+            fir_options: self.fir_options,
+            thread_pool: self.thread_pool.clone(),
+        }
     }
 
     /// get_algorithm() -> Tuple[int, int, int]
@@ -53,7 +63,7 @@ impl RustResizeOptions {
     ///
     /// :rtype: Tuple[int, int, int]
     fn get_resize_alg(&self) -> (u8, u8, u8) {
-        let (algorithm, filter_type, multiplicity) = match self.0.algorithm {
+        let (algorithm, filter_type, multiplicity) = match self.fir_options.algorithm {
             fr::ResizeAlg::Nearest => (1u8, 0u8, 2u8),
             fr::ResizeAlg::Convolution(filter_type) => (2u8, filter_type_as_u8(filter_type), 2u8),
             fr::ResizeAlg::SuperSampling(filter_type, multiplicity) => {
@@ -61,12 +71,6 @@ impl RustResizeOptions {
             }
             _ => (0u8, 0u8, 2u8),
         };
-
-        // let algorithm = algorithm.to_object(py);
-        // let filter_type = filter_type.to_object(py);
-        // let multiplicity = multiplicity.to_object(py);
-
-        // let res: PyObject = PyTuple::new_bound(py, &[algorithm, filter_type, multiplicity]).into();
         (algorithm, filter_type, multiplicity)
     }
 
@@ -79,13 +83,16 @@ impl RustResizeOptions {
             3 => fr::ResizeAlg::SuperSampling(filter_type_from_u8(filter_type), multiplicity),
             _ => fr::ResizeAlg::Nearest,
         };
-        Self(self.0.resize_alg(resizer_alg))
+        Self {
+            fir_options: self.fir_options.resize_alg(resizer_alg),
+            thread_pool: self.thread_pool.clone(),
+        }
     }
 
     /// Set crop box for source image.
     fn get_crop_box(&self) -> Option<(f64, f64, f64, f64)> {
-        match self.0.cropping {
-            SrcCropping::Crop(crop_box) => {
+        match self.fir_options.cropping {
+            fr::SrcCropping::Crop(crop_box) => {
                 Some((crop_box.left, crop_box.top, crop_box.width, crop_box.height))
             }
             _ => None,
@@ -95,12 +102,15 @@ impl RustResizeOptions {
     /// Set crop box for source image.
     #[pyo3(signature = (left, top, width, height))]
     fn set_crop_box(&self, left: f64, top: f64, width: f64, height: f64) -> Self {
-        Self(self.0.crop(left, top, width, height))
+        Self {
+            fir_options: self.fir_options.crop(left, top, width, height),
+            thread_pool: self.thread_pool.clone(),
+        }
     }
 
     fn get_fit_into_destination_centering(&self) -> Option<(f64, f64)> {
-        match self.0.cropping {
-            SrcCropping::FitIntoDestination(centering) => Some(centering),
+        match self.fir_options.cropping {
+            fr::SrcCropping::FitIntoDestination(centering) => Some(centering),
             _ => None,
         }
     }
@@ -108,17 +118,42 @@ impl RustResizeOptions {
     /// Fit source image into the aspect ratio of destination image without distortions.
     #[pyo3(signature = (centering=None))]
     fn set_fit_into_destination(&self, centering: Option<(f64, f64)>) -> Self {
-        Self(self.0.fit_into_destination(centering))
+        Self {
+            fir_options: self.fir_options.fit_into_destination(centering),
+            thread_pool: self.thread_pool.clone(),
+        }
     }
 
     fn get_use_alpha(&self) -> bool {
-        self.0.mul_div_alpha
+        self.fir_options.mul_div_alpha
     }
 
     /// Enable or disable consideration of the alpha channel when resizing.
     #[pyo3(signature = (v))]
     fn set_use_alpha(&self, v: bool) -> Self {
-        Self(self.0.use_alpha(v))
+        Self {
+            fir_options: self.fir_options.use_alpha(v),
+            thread_pool: self.thread_pool.clone(),
+        }
+    }
+
+    /// get_thread_pool() -> Option<ThreadPool>
+    /// --
+    ///
+    /// Returns thread pool.
+    ///
+    /// :rtype: Option<ThreadPool>
+    fn get_thread_pool(&self) -> Option<ResizerThreadPool> {
+        self.thread_pool.clone()
+    }
+
+    /// Set thread pool.
+    #[pyo3(signature = (thread_pool))]
+    fn set_thread_pool(&mut self, thread_pool: Option<ResizerThreadPool>) -> Self {
+        Self {
+            fir_options: self.fir_options,
+            thread_pool,
+        }
     }
 }
 
@@ -138,11 +173,11 @@ impl RustResizer {
 
     /// Returns CPU extensions.
     #[pyo3(text_signature = "($self)")]
-    fn get_cpu_extensions(&self, py: Python) -> PyResult<PyObject> {
+    fn get_cpu_extensions<'py>(&self, py: Python<'py>) -> PyResult<Bound<'py, PyInt>> {
         let resizer_mutex = self.resizer.clone();
         let resizer = result2pyresult(resizer_mutex.lock())?;
         let cpu_extensions = cpu_extensions_to_u8(resizer.cpu_extensions());
-        Ok(cpu_extensions.into_py(py))
+        Ok(cpu_extensions.into_pyobject(py)?)
     }
 
     /// Set CPU extensions.
@@ -168,12 +203,23 @@ impl RustResizer {
     ) -> PyResult<()> {
         let resizer_mutex = self.resizer.clone();
         py.allow_threads(move || {
-            let fir_options = options.map(|o| &o.0);
+            let fir_options = options.map(|o| &o.fir_options);
             let src_image_view = src_image.src_image_view();
             let dst_image_view = dst_image.dst_image_view();
-            let mut resizer = result2pyresult(resizer_mutex.lock())?;
-            result2pyresult(resizer.resize(src_image_view, dst_image_view, fir_options))?;
-            Ok(())
+            let mut resizer_guard = result2pyresult(resizer_mutex.lock())?;
+            let resizer = resizer_guard.deref_mut();
+            if let Some(thread_pool) = options.and_then(|o| o.thread_pool.as_ref()) {
+                if fir_options
+                    // Don't process in thread-pool if resize alg is nearest
+                    .map(|o| o.algorithm != fr::ResizeAlg::Nearest)
+                    .unwrap_or(true)
+                {
+                    return thread_pool.run_within(|| {
+                        result2pyresult(resizer.resize(src_image_view, dst_image_view, fir_options))
+                    });
+                }
+            }
+            result2pyresult(resizer.resize(src_image_view, dst_image_view, fir_options))
         })
     }
 
@@ -188,10 +234,21 @@ impl RustResizer {
     ) -> PyResult<()> {
         let resizer_mutex = self.resizer.clone();
         py.allow_threads(move || {
-            let fir_options = options.map(|o| &o.0);
-            let mut resizer = result2pyresult(resizer_mutex.lock())?;
-            result2pyresult(resizer.resize(src_image, dst_image, fir_options))?;
-            Ok(())
+            let fir_options = options.map(|o| &o.fir_options);
+            let mut resizer_guard = result2pyresult(resizer_mutex.lock())?;
+            let resizer = resizer_guard.deref_mut();
+            if let Some(thread_pool) = options.and_then(|o| o.thread_pool.as_ref()) {
+                if fir_options
+                    // Don't process in thread-pool if resize alg is nearest
+                    .map(|o| o.algorithm != fr::ResizeAlg::Nearest)
+                    .unwrap_or(true)
+                {
+                    return thread_pool.run_within(|| {
+                        result2pyresult(resizer.resize(src_image, dst_image, fir_options))
+                    });
+                }
+            }
+            result2pyresult(resizer.resize(src_image, dst_image, fir_options))
         })
     }
 }